requirements.txt
requests
beautifulsoup4
kafka-python
pandas